// src/types/staticsCourse.ts
export type StatisticType = {
  percentage: number;
  completed_lessons: number;
  amount_lessons: number;
  completed_videos: number;
  amount_videos: number;
  received_points: number;
  amount_points: number;
  completed_tests: number;
  amount_tests: number;
  completed_questions: number;
  amount_questions: number;
};